# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MediaItem {
  id: number;
  title?: string;
  name?: string;
  overview?: string;
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  vote_average?: number;
  vote_count?: number;
  release_date?: string;
  first_air_date?: string;
  media_type?: string;
  genre_ids?: number[];
  popularity?: number;
}

export interface SearchResults {
  results: MediaItem[];
  total_results: number;
  total_pages: number;
  page: number;
}

export interface Genre {
  id: number;
  name: string;
}

export interface GenreList {
  movie_genres: Genre[];
  tv_genres: Genre[];
}

export interface Season {
  id: number;
  season_number: number;
  name?: string;
  episode_count?: number;
  /** @nullable */
  poster_path?: string | null;
  air_date?: string;
}

export interface MediaDetails {
  id: number;
  title?: string;
  name?: string;
  overview?: string;
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  vote_average?: number;
  vote_count?: number;
  release_date?: string;
  first_air_date?: string;
  /** @nullable */
  runtime?: number | null;
  status?: string;
  tagline?: string;
  genres?: Genre[];
  number_of_seasons?: number;
  number_of_episodes?: number;
  seasons?: Season[];
  media_type?: string;
}

export interface Episode {
  id: number;
  episode_number: number;
  name?: string;
  overview?: string;
  /** @nullable */
  still_path?: string | null;
  air_date?: string;
  /** @nullable */
  runtime?: number | null;
  vote_average?: number;
}

export interface SeasonDetails {
  id: number;
  season_number: number;
  name?: string;
  overview?: string;
  /** @nullable */
  poster_path?: string | null;
  air_date?: string;
  episodes: Episode[];
}

export interface CastMember {
  id: number;
  name: string;
  character?: string;
  /** @nullable */
  profile_path?: string | null;
  order?: number;
}

export interface Credits {
  cast: CastMember[];
  crew: CastMember[];
}

export interface VideoItem {
  id: string;
  key: string;
  name: string;
  site: string;
  type: string;
  official?: boolean;
  published_at?: string;
}

export interface Videos {
  results: VideoItem[];
}

export type SearchMoviesParams = {
query: string;
page?: number;
mediaType?: SearchMoviesMediaType;
};

export type SearchMoviesMediaType = typeof SearchMoviesMediaType[keyof typeof SearchMoviesMediaType];


export const SearchMoviesMediaType = {
  movie: 'movie',
  tv: 'tv',
  all: 'all',
} as const;

export type GetTrendingParams = {
mediaType?: GetTrendingMediaType;
timeWindow?: GetTrendingTimeWindow;
};

export type GetTrendingMediaType = typeof GetTrendingMediaType[keyof typeof GetTrendingMediaType];


export const GetTrendingMediaType = {
  movie: 'movie',
  tv: 'tv',
  all: 'all',
} as const;

export type GetTrendingTimeWindow = typeof GetTrendingTimeWindow[keyof typeof GetTrendingTimeWindow];


export const GetTrendingTimeWindow = {
  day: 'day',
  week: 'week',
} as const;

export type GetPopularParams = {
page?: number;
};

export type GetPopularTvParams = {
page?: number;
};

export type GetTopRatedParams = {
page?: number;
};

export type GetNowPlayingParams = {
page?: number;
};